dagster_home/pyproject.toml
[project]
name = "local_dagster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
    "dagster",
    "dagster-cloud",
    "dagster-dbt",
    "dbt-duckdb<1.10",
]

[project.optional-dependencies]
dev = [
    "dagster-webserver",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.dagster]
module_name = "local_dagster.definitions"
code_location_name = "local_dagster"
